fix(compile errors): fixed wide range of compile errors
Added eslint and prettier for formating and linting. Removed body-parser as it was redundant. Fixed
a good chunk of the compile errors, but a great deal still remain. Replaced require statements with
imports, if they were not entirely removed. Added more sentry exception capturing to try catch
blocks. Reworked some logger imports to import from the utils index file and not the logger util
file.